requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.14
django-allauth==0.39.1
django-cors-headers==3.0.2
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-environ==0.4.5
django-templated-mail==1.1.1
djangorestframework==3.9.4
djoser==2.0.3
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
Pillow==6.2.2
pycodestyle==2.5.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
six==1.15.0
sqlparse==0.2.4
stripe==2.27.0
typed-ast==1.4.0
urllib3==1.24.2
whitenoise==5.2.0
wrapt==1.11.2
